hhkl/requirements.txt
Django==1.8.7
Fabric==1.10.2
django-debug-toolbar==1.4
gunicorn==19.4.1
paramiko==1.16
psycopg2==2.6.1
pycrypto==2.6.1
django-tastypie==0.12.2
